Move Pandoc subprocess call to a separate thread
This commit moves the Pandoc subprocess call into a new thread - the
benefit of this is we can now open a log file and use it to redirect
stdout and stderr into, then properly close it once the Pandoc call is
complete. Previously we started the subprocess but could not wait for it
to return because CherryPy needs to reply to the client request.