Add backend support for latexdiff
Refs #6
This commit introduces a number of changes to accept and use the
"compare mode" implementation of the md2pdf-client. It firstly looks for
the relevant HTTP header, and if found then looks for MD files with
".old" and ".new" in the filenames. It takes these, runs it through a
string replace filter (in case different options are desired for the
compared version, e.g. TOC doesn't work very well, so it can be removed)
and then finally it creates two LaTeX files, runs latexdiff, and
produces the PDF for return to the client.