Fix latexdiff not working with hypertarget
Refs #6
Previously, the PDF generated by the compare mode did not highlight
changes in headings. This turned out to be due to the way Pandoc uses
hyperref to generate hyperlink targets - and this added some LaTeX code
to the source that confused latexdiff. This commit adds an extra pre-
processing step to strip out the hypertarget references, which then
fixes the output PDF.