Change chroot installer to use TeX Live ISO
Using the online install method for TeX Live sometimes poses issues, for
example if the Internet connection is poor the installer may crash, and
so far no resume support has been added to md2pdf. There is also the
problem of consistency - a user may want to intentionally use an older
TeX Live release.
This commit changes the chroot install method to use the ISO image of
TeX Live. This must be manually downloaded by the user.
To mount the ISO inside the Snap package, the 'fuse-control' slot is
required - this needs to be manually connected.