Use bindfs for /dev
Previously, the installer set up static files inside the chroot for
/dev/random and /dev/urandom - this was because the chroot is unable to
access the real devices on the host. Typically when using a chroot this
would be dealt with using a 'bind' mount - but this cannot be done
inside a Snap package, so a static file was used.
This commit changes the chroot setup to instead use a FUSE bind mount.